requirements.txt
asgiref==3.11.1
certifi==2026.4.22
charset-normalizer==3.4.7
Django==6.0.5
django-cors-headers==4.9.0
django-unfold==0.93.0
djangorestframework==3.17.1
et_xmlfile==2.0.0
idna==3.15
openpyxl==3.1.5
pillow==12.2.0
python-decouple==3.8
requests==2.34.1
sqlparse==0.5.5
urllib3==2.7.0
gunicorn==23.0.0
whitenoise==6.9.0
